+++ dataquality-common/CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
-N/A
+- feat(TDQ-18062): Add StringChecker, StringHandler, TokenizedString and Acronym utility classes in new "character" package
 ### Changed
 N/A
 ### Removed